requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
transformers
nltk
python-dotenv
streamlit
numpy
tokenizers 